Game page has been refactored
