+++ client/package.json
@@ -52,5 +52,5 @@
     "typescript-eslint": "^8.7.0",
     "vite": "^5.4.7"
   },
-  "packageManager": "pnpm@10.4.1+sha512.c753b6c3ad7afa13af388fa6d808035a008e30ea9993f58c6663e2bc5ff21679aa834db094987129aa4d488b86df57f7b634981b2f827cdcacc698cc0cfb88af"
+  "packageManager": "pnpm@10.6.5"
 }
